+++ client/src/components/TodoList.tsx
@@ -1,23 +1,19 @@
 import React, { useState } from 'react';
-
 import { useFragment } from "react-relay/hooks"
 import graphql from 'babel-plugin-relay/macro';
 import TodoListItem from './TodoListItem';
-import TextField from '@mui/material/TextField';
 import { FormGroup } from '@mui/material';
 import { TodoList_query$key } from './__generated__/TodoList_query.graphql';
+import CreateTodo from './CreateTodo';
+
 
 type Props = {
     query: TodoList_query$key,
-    addTodo: any,
+    
 }
 
 
-// TODO: proper types
 export default function TodoList(props: Props){
-
-    const [input, setInput] = useState('');
-
     const data = useFragment(
         graphql`
         fragment TodoList_query on Query {
@@ -31,37 +27,19 @@ export default function TodoList(props: Props){
     `, 
     props.query
     );
-
-    function handleKeyDown (e: React.KeyboardEvent<HTMLInputElement>) {
-        console.log(e.key)
-        if(e.key === 'Enter'){
-            props.addTodo(input)
-            setInput('')
-        }
-    }
-
-    function handleInputChange (e: React.ChangeEvent<HTMLInputElement>) {
-        setInput(e.currentTarget.value);
-    }
     
-    // TODO: ensure allTodos exists on data object 
-    const list = data.allTodos!.nodes.map((todo: any) => (
-      <TodoListItem
-        key={todo.id}
-        todo={todo}
-        />
-    ))
     return (
         <>
-            <TextField 
-                id="todo-input" 
-                label="Todo" 
-                variant="outlined" 
-                onChange={handleInputChange} 
-                onKeyDown={handleKeyDown} 
-                value={input}/>
+            <CreateTodo />
             <FormGroup>
-                { list }
+                {data.allTodos && 
+                    data.allTodos.nodes.map((todo: any) => (
+                        <TodoListItem
+                          key={todo.id}
+                          todo={todo}
+                          />
+                      ))  
+                }
             </FormGroup>
         </>
