+++ client/src/components/TodoListItem.tsx
@@ -1,21 +1,22 @@
 
 import RelayEnvironment from '../RelayEnvironment';
-import { useFragment } from 'react-relay/hooks';
+import { useFragment, useMutation } from 'react-relay/hooks';
 import graphql from 'babel-plugin-relay/macro';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import Checkbox from '@mui/material/Checkbox';
-import React, { useState } from "react";
+import React from "react";
 import { TodoListItem_todo$key } from './__generated__/TodoListItem_todo.graphql';
 import { Button } from '@mui/material';
 
 import { commitChangeTodoMutation } from '../mutations/ChangeTodoMutation';
 import { commitDeleteTodoMutation } from '../mutations/DeleteTodoMutation';
 
-type Props = {
-    todo: TodoListItem_todo$key
-}
+interface Props {
+    todo: TodoListItem_todo$key;
+  }
 
-export default function TodoListItem(props: Props){
+
+export default function TodoListItem({ todo }: Props){
     const data = useFragment(
         graphql`
             fragment TodoListItem_todo on Todo {
@@ -25,24 +26,23 @@ export default function TodoListItem(props: Props){
                 completed
             }
         `,
-        props.todo,
+        todo,
     );
 
-    const [checked, setChecked] = useState(data.completed)
 
-    const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const handleCompleteChange = (event: React.ChangeEvent<HTMLInputElement>) => {
         commitChangeTodoMutation(
             RelayEnvironment, 
             data.id, 
             !data.completed, 
             () => {
                 console.log('change todo mutation successful');
-                setChecked(!data.completed);
             }
         )
     }
 
     const handleDelete = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
+        console.log(event);
         commitDeleteTodoMutation(
             RelayEnvironment,
             data.rowId,
@@ -57,8 +57,8 @@ export default function TodoListItem(props: Props){
                     <td>
                     <FormControlLabel 
                         control={<Checkbox 
-                            checked={checked}
-                            onChange={handleChange} 
+                            checked={data.completed}
+                            onChange={handleCompleteChange} 
                             />} 
                         label={data.task} 
                     />
